Tests/LibWeb/Text/expected/textarea-placeholder-shown.txt

empty value: true
non-empty value: false
no placeholder attr: false
empty placeholder attr: true
value cleared: true
placeholder removed: false
